ict_trading_bot/requirements.txt
# Core Trading
MetaTrader5>=5.0.45
python-mt5-connector>=1.0.0

# Data Processing
pandas>=1.5.3
numpy>=1.24.0

# Analysis & ML
scikit-learn>=1.2.0
xgboost>=1.7.0
ta>=0.10.2

# Web & API
Flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# Real-time & Websockets
ws>=4.0.0
websockets>=11.0.0

# Data & Config
pytz>=2023.3
PyYAML>=6.0

# Database
supabase>=2.0.0
postgrest>=0.10.0

# Utilities
python-dateutil>=2.8.2
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
